Extract shared PhotoSwipe viewer core
