+++ web/apps/albums/src/pages/_app.tsx
@@ -3,10 +3,7 @@ import type { AppProps } from "next/app";
 
 import "@/app/styles/fonts.css";
 import "@/app/styles/global.css";
-import "photoswipe/dist/photoswipe.css";
-// Keep our PhotoSwipe overrides after the library CSS so the reduced z-index
-// wins over PhotoSwipe's default 10k layer.
-import "@/app/styles/photoswipe.css";
+import "ente-gallery/styles/photoswipe.css";
 
 type AlbumsAppProps = AppProps<Record<string, unknown>>;
