+++ web/apps/photos/src/pages/_app.tsx
@@ -56,9 +56,8 @@ import type { AppProps } from "next/app";
 import { useRouter } from "next/router";
 import { useCallback, useEffect, useMemo, useState } from "react";
 
-import "photoswipe/dist/photoswipe.css";
+import "ente-gallery/styles/photoswipe.css";
 import "../styles/global.css";
-import "../styles/photoswipe.css";
 
 type PhotosAppProps = AppProps<Record<string, unknown>>;
